+++ crates/client/mempool/src/close_block.rs
@@ -1,21 +1,24 @@
-use dc_db::DeoxysBackend;
-use dc_sync::{
-    commitments::update_tries_and_compute_state_root,
-    convert::{compute_commitments_for_block, BlockCommitments},
+use dc_block_import::{
+    BlockImportError, BlockImportResult, BlockImporter, UnverifiedFullBlock, UnverifiedHeader, Validation,
 };
-use dp_block::{
-    header::PendingHeader, DeoxysBlock, DeoxysBlockInfo, DeoxysPendingBlock, DeoxysPendingBlockInfo, Header,
-};
+use dp_block::{header::PendingHeader, DeoxysPendingBlock, DeoxysPendingBlockInfo};
 use dp_state_update::StateDiff;
-use starknet_types_core::felt::Felt;
+use starknet_api::core::ChainId;
 
-pub fn close_block(
-    backend: &DeoxysBackend,
+/// Close the block (convert from pending to closed), and store to db. This is delegated to the block import module.
+pub async fn close_block(
+    importer: &BlockImporter,
     block: DeoxysPendingBlock,
     state_diff: &StateDiff,
-    chain_id: Felt,
+    chain_id: ChainId,
     block_number: u64,
-) -> DeoxysBlock {
+) -> Result<BlockImportResult, BlockImportError> {
+    let validation = Validation {
+        trust_transaction_hashes: true, // no need to recompute tx hashes
+        chain_id,
+        trust_global_tries: false,
+    };
+
     let DeoxysPendingBlock { info, inner } = block;
     let DeoxysPendingBlockInfo { header, tx_hashes: _tx_hashes } = info;
 
@@ -29,45 +32,27 @@ pub fn close_block(
         l1_da_mode,
     } = header;
 
-    let (global_state_root, block_commitments) = rayon::join(
-        || update_tries_and_compute_state_root(backend, state_diff, block_number),
-        || compute_commitments_for_block(&inner, state_diff, protocol_version, chain_id, block_number),
-    );
+    let block = importer
+        .pre_validate(
+            UnverifiedFullBlock {
+                unverified_block_number: Some(block_number),
+                header: UnverifiedHeader {
+                    parent_block_hash: Some(parent_block_hash),
+                    sequencer_address,
+                    block_timestamp,
+                    protocol_version,
+                    l1_gas_price,
+                    l1_da_mode,
+                },
+                state_diff: state_diff.clone(),
+                transactions: inner.transactions,
+                receipts: inner.receipts,
+                declared_classes: vec![],
+                commitments: Default::default(), // the block importer will compute the commitments for us
+            },
+            validation.clone(),
+        )
+        .await?;
 
-    let BlockCommitments {
-        transaction_commitment,
-        transaction_count,
-        event_commitment,
-        event_count,
-        receipt_commitment,
-        state_diff_commitment,
-        state_diff_length,
-        tx_hashes,
-    } = block_commitments;
-
-    let header = Header {
-        parent_block_hash,
-        sequencer_address,
-        block_timestamp,
-        protocol_version,
-        l1_gas_price,
-        l1_da_mode,
-
-        // Extra fields.
-        block_number,
-
-        // Commitments.
-        global_state_root,
-        transaction_count,
-        transaction_commitment,
-        event_count,
-        event_commitment,
-        state_diff_length,
-        state_diff_commitment,
-        receipt_commitment,
-    };
-
-    let block_hash = header.compute_hash(chain_id);
-
-    DeoxysBlock { info: DeoxysBlockInfo { header, block_hash, tx_hashes }, inner }
+    importer.verify_apply(block, validation.clone()).await
 }